backend/services/api/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
asyncpg==0.29.0
paho-mqtt==2.0.0
websockets==12.0
